Fixed update script for BE

+++ server/install.ps1
@@ -4,8 +4,11 @@
 $REQUREMENTS_LOCK_FILE = 'requirements.txt'
 
 'Installing server...' | Out-Host
-"Installing requirements from requirements-lock file $REQUREMENTS_LOCK_FILE..." | Out-Host
-& pip install -r $REQUREMENTS_LOCK_FILE
+
+if (Test-Path -PathType Leaf -Path (Join-Path '.' $REQUREMENTS_LOCK_FILE)) {
+    "Installing requirements from requirements-lock file $REQUREMENTS_LOCK_FILE..." | Out-Host
+    & pip install -r $REQUREMENTS_LOCK_FILE
+}
 
 Get-ChildItem -Path (Join-Path '.' '*requirements.txt') -Exclude $REQUREMENTS_LOCK_FILE | ForEach-Object {
     "Installing requirements from $($_.Name)..." | Out-Host
